requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.6
djangorestframework==3.15.1
et-xmlfile==1.1.0
mypy-extensions==1.0.0
openpyxl==3.1.2
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
uWSGI==2.0.25.1
